requirements.txt
dataset==1.5.2
fastapi==0.75.2
happytransformer==2.4.1
pandas==1.3.4
pandasql==0.7.3
pydantic==1.8.2
pyocclient==0.6
requests==2.26.0
starlette==0.17.1
streamlit==1.8.1
tqdm==4.62.3
uvicorn==0.17.6
nltk
word2number
scispacy